blog_new/requirements.txt
asn1crypto==0.24.0
bootstrap-admin==0.4.3
cachetools==2.1.0
certifi==2018.4.16
cffi==1.13.2
chardet==3.0.4
cryptography==2.3
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-appconf==1.0.2
django-braces==1.13.0
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-haystack==2.8.1
django-imagekit==4.0.2
django-jet==1.0.7
django-js-asset==1.1.0
django-mdeditor==0.1.17
django-model-utils==3.1.2
django-mptt==0.9.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-simpleui==3.9.1
django-smtp-ssl==1.0
django-taggit==0.22.2
django-uuslug==1.1.8
django-widget-tweaks==1.4.5
djangorestframework==3.8.2
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
jieba==0.39
jsonfield==2.0.2
Markdown==2.6.11
oauth2client==4.1.2
oauthlib==2.1.0
payjs==1.2.5
pilkit==2.0
Pillow==7.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.3.1
PyMySQL==0.9.2
python-slugify==1.2.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
setuptools-scm==3.5.0
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.2
tornado==6.0.4
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
Whoosh==2.7.4
